bot-ia/requirements.txt
llama-cpp-python[server]
huggingface-hub
uvicorn
requests
